requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
Pillow>=10.1.0
django-filter>=23.5
djangorestframework-simplejwt>=5.3.0
social-auth-app-django>=5.4.0
drf-spectacular>=0.27.0
celery>=5.3.4
redis>=5.0.1
django-celery-beat>=2.5.0
dj-database-url>=2.1.0
whitenoise>=6.6.0
gunicorn>=21.2.0
google-auth>=2.25.0
djangorestframework-simplejwt[crypto]>=5.3.0
icalendar>=5.0.11

# Testing
factory-boy>=3.3.0
Faker>=21.0.0
pytest>=7.4.3
pytest-django>=4.7.0
pytest-cov>=4.1.0
